shows submit button in awaiting current user signature panel if can write and no missing signatures
